Refactored the text generation methods for models to the model definition file
