+++ examples/t5.rs
@@ -12,9 +12,9 @@
 
 extern crate anyhow;
 
-use rust_bert::pipelines::generation_utils::{GenerateConfig, LanguageGenerator, T5Generator};
+use rust_bert::pipelines::generation_utils::{GenerateConfig, LanguageGenerator};
 use rust_bert::resources::{RemoteResource, Resource};
-use rust_bert::t5::{T5ConfigResources, T5ModelResources, T5VocabResources};
+use rust_bert::t5::{T5ConfigResources, T5Generator, T5ModelResources, T5VocabResources};
 
 fn main() -> anyhow::Result<()> {
     //    Resources paths
