+++ src/openai_gpt/mod.rs
@@ -59,6 +59,6 @@ mod openai_gpt_model;
 mod transformer;
 
 pub use openai_gpt_model::{
-    OpenAIGPTLMHeadModel, OpenAiGptConfigResources, OpenAiGptMergesResources, OpenAiGptModel,
-    OpenAiGptModelOutput, OpenAiGptModelResources, OpenAiGptVocabResources,
+    OpenAIGPTLMHeadModel, OpenAIGenerator, OpenAiGptConfigResources, OpenAiGptMergesResources,
+    OpenAiGptModel, OpenAiGptModelOutput, OpenAiGptModelResources, OpenAiGptVocabResources,
 };
