+++ src/t5/mod.rs
@@ -55,6 +55,6 @@ mod t5_model;
 
 pub use attention::LayerState;
 pub use t5_model::{
-    T5Config, T5ConfigResources, T5ForConditionalGeneration, T5Model, T5ModelOutput,
+    T5Config, T5ConfigResources, T5ForConditionalGeneration, T5Generator, T5Model, T5ModelOutput,
     T5ModelResources, T5Prefix, T5VocabResources,
 };
